LC-237: Small cleanup / test updates / more JavaDoc
