+++ lucille-core/src/test/java/com/kmwllc/lucille/core/RunnerTest.java
@@ -823,4 +823,22 @@ public class RunnerTest {
     Config config = ConfigFactory.load("RunnerTest/render.conf");
     Runner.renderConfig(config);
   }
+
+  @Test
+  public void testStringifyValidationExceptions() {
+    Map<String, List<Exception>> exceptions = new LinkedHashMap<>();
+    exceptions.put("pipeline1", List.of(new Exception("exception 1"), new Exception("exception 2")));
+    exceptions.put("pipeline2", List.of(new Exception("exception 3")));
+
+    String expected = "Pipeline Configuration is invalid. Printing the list of exceptions for each pipeline\n"
+        + "\tPipeline: pipeline1\tError count: 2\n\t\tException 1: exception 1\n\t\tException 2: exception 2\n"
+        + "\tPipeline: pipeline2\tError count: 1\n\t\tException 1: exception 3";
+    assertEquals(expected, Runner.stringifyValidation(exceptions, "Pipeline"));
+  }
+
+  @Test
+  public void testStringifyValidationNoExceptions() {
+    Map<String, List<Exception>> exceptions = new LinkedHashMap<>();
+    assertEquals("Pipeline Configuration is valid", Runner.stringifyValidation(exceptions, "Pipeline"));
+  }
 }
